fix(comments): review comments incorporated
